Bug fixes, audio, button class, TEsound
got the game to a more or less bug-free state.

added audio. game now plays Clair De Lune in the background, plays a
clicking noise when placing a piece, and plays an airhorn sound when a
player wins.

I also started working on mute buttons for the music/sfx (they don’t
work yet), and the class i am working on for these buttons will be very
useful throughout the rest of the interface as we add more features in
the future.

TEsound.lua is an audio management system for Love 2d that I found.
I’ve yet to do much with it, but i plan on trying it out soon.

I haven’t really done anything with the “cell” class yet.